refactor: move profitable and delayed status into hook
